Abstracted Party and sub-components to it's own file. Now a little beautification.
